Renamed ffmpeg.py so it doesnt conflict with the ffmpeg command
Removed the .exe from ffmpeg commands and renamed ffmpeg.py to pyffmpeg.py so that the program will still work on linux based systems and not try to execute the ffmpeg python file
